unsquashfs: Preserve times on symlink inodes
This patch fixes a bug that caused symlinks created by unsquashfs to
use the current time for atime and mtime rather than the times present
in the squashfs archive.

The timestamps are set using utimensat(2) which does not dereference the
symlink when the AT_SYMLINK_NOFOLLOW flag is used. The utimensat(2)
system call supports nanosecond precision but that's not needed by
squashfs so timespec.tv_nsec is unconditionally set to 0.

Signed-off-by: Tyler Hicks <tyhicks@canonical.com>
Bug-Ubuntu: https://launchpad.net/bugs/1724152
Forwarded: https://sourceforge.net/p/squashfs/mailman/message/36140758/